+++ lib/src/botapi/filters.dart
@@ -2,20 +2,27 @@ import 'package:vklib/src/core/longpoll/group_lp_objects/message_new.dart';
 
 typedef filterType<T> = T Function(MessageNewObject obj);
 
+class BaseFilter {
+  BaseFilter(this.filter);
+
+  filterType filter;
+
+  Future<bool> execute(MessageNewObject obj) async {
+    return filter(obj) as bool;
+  }
+}
+
+/// Standart Kit of Filters
 class BotFilters {
-  static filterType ignoreBots() {
-    return (MessageNewObject obj) {
-      return obj.isUser;
-    };
+  static BaseFilter ignoreBots() {
+    return BaseFilter((obj) => obj.isUser);
   }
 
-  static filterType chatOnly() {
-    return (MessageNewObject obj) {
-      return obj.isChat;
-    };
+  static BaseFilter chatOnly() {
+    return BaseFilter((obj) => obj.isChat);
   }
 
-  static filterType other(filterType _filter) {
-    return _filter;
+  static BaseFilter other(filterType _filter) {
+    return BaseFilter(_filter);
   }
 }