+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.4.0 botAPI
+
+* add router for bot command
+* add bot command extends RegExp, justText, command
+* add command filters
+* add typing argument
+* add messageNewContext
+
 ## 0.3.0
 
 * add message context for GroupLongPoll